Fix up bug in local media player state restoration.
Once state was restored, the onPlay callback was being
invoked, but this was incorrect because it assumes that
the media player was already set up.

Changing the behavior of this method would break the
play/pause functionality so instead put the necessary code
inline, it is not a lot of code and it keeps things simple.

Bug: 32097748
Test: Saved and restored playlist successfully.
Change-Id: Id442270b5bd13c5c5126464a6788a82465711391