Make LocalMediaPlayer post a notification.
We need local media player to post a notification now since we don't
automatically create ongoing action cards for media anymore given
that we replaced home with the notification shade.

While it would be convenient to post the notification from the media
player, it would cause problems with 3rd party applications who post
their own notifications, so it's best to post the notification from
local media player itself.

Having a notification also keeps the service running in the background.

Test: Manually verified.
Change-Id: Ie63642ca616ad3cb931b07fd93c97125f4b0574c