Use correct list for song shuffling.
I incorrectly assumed that the DataModel class held the source to truth for
the current state to the play queue. However it holds the results of the
last query done from LocalMeidaBrowserService. Thus it is not in the same
state as that of the MediaSession queue which is what we are actually
shuffling. Once again I'm just patching the errors in this class until a
refactor can be planned.

Bug: 65161162

Test: After starting to play music browser for a new song but don't select
      it. Then shuffle the songs again and confirm the same list was shuffled

Change-Id: I96df1436fe3674a3d4a07b8b907194aed0caf31b