Add required flag to registerReceiver call in LocalMediaBrowserService
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_NOT_EXPORTED
flag to the call to registerReceiver in LocalMediaBrowserService when
registering for the media control actions as these are only broadcast
as PendingIntents with the calling identity of the local app.

Bug: 161145287
Test: Build
Change-Id: I3973516514593f26abfaeb0d26ead7a8fe3d4845